Created matlab script that integrates with VRep
Includes setup, parameters about the cart, Vrep remote api connection configuration, and object handles for everything. Includes 3 functions to control the cart and arm. The first (Drive) takes in a distance and speed and drives the cart forward or reverse. The second rotates the cart about it's center. The third takes an end effector pose as its input, computes the inverse kinematics, and moves the robot arm to that pose.